# Eaas/openapi.yaml
openapi: 3.0.3
info:
  title: EaaS — Escrow as a Service API
  description: |
    API-first, Africa-native escrow platform. Enables secure transactions
    between buyers and sellers by holding funds in escrow until delivery is confirmed.
    
    ## Authentication
    - **Customer/Admin:** Bearer JWT token via `Authorization: Bearer <token>`
    - **Merchant API:** API Key via `X-API-Key: sk_live_xxxx` header
    - **Webhooks from PSP:** Verified via `X-Interswitch-Signature` HMAC header
    
    ## Idempotency
    All POST requests that create resources must include `X-Idempotency-Key` header.
    
    ## Rate Limiting
    - 100 requests/minute per API key
    - Rate limit headers returned: `X-RateLimit-Limit`, `X-RateLimit-Remaining`
    
  version: 1.0.0
  contact:
    name: EaaS Developer Support
    email: developers@eaas.africa
servers:
  - url: https://api.eaas.africa/api/v1
    description: Production
  - url: https://sandbox.eaas.africa/api/v1
    description: Sandbox
...
paths:
...
  /payouts:
    get:
      tags: [Payouts]
      summary: List my payouts
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Payout list
  /payouts/{id}:
    get:
      tags: [Payouts]
      summary: Get payout details
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Payout details

  # ─── ADMIN ─────────────────────────────────────────────────
...
  /admin/payouts:
    get:
      tags: [Admin]
      summary: View all payouts and their status
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [PENDING, PROCESSING, SUCCESS, FAILED]
      responses:
        '200':
          description: Payout queue
...
  /admin/aml-alerts:
    get:
      tags: [AML / Compliance]
      summary: Admin list AML alerts
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of AML alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AmlAlert'
                  totalElements:
                    type: integer
                  totalPages:
                    type: integer
                  page:
                    type: integer
                  size:
                    type: integer
  /admin/aml-alerts/{id}/review:
    put:
      tags: [AML / Compliance]
      summary: Review alert
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: notes
          schema:
            type: string
        - in: header
          name: X-User-Id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert reviewed
        '404':
          $ref: '#/components/responses/NotFound'
  /admin/aml-alerts/{id}/resolve:
    put:
      tags: [AML / Compliance]
      summary: Resolve alert
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
        - in: query
          name: resolution
          schema:
            type: string
        - in: query
          name: notes
          schema:
            type: string
      responses:
        '200':
          description: Alert resolved
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          example:
            status: "error"
            message: "Merchant not found"
...
  schemas:
...
    AmlAlert:
      type: object
      properties:
        id:
          type: string
          format: uuid
        paymentId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        alertType:
          type: string
          example: "THRESHOLD_EXCEEDED"
        amount:
          type: number
          format: decimal
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, CLOSED_FALSE_POSITIVE, CONFIRMED_SUSPICIOUS, SAR_FILED]
        notes:
          type: string
        createdAt:
          type: string
          format: date-time
...